requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
boto3
python-dotenv
python-multipart